+++ packages/video_player/video_player/lib/video_player.dart
@@ -32,6 +32,7 @@ class VideoPlayerValue {
     this.isPlaying = false,
     this.isLooping = false,
     this.isBuffering = false,
+    this.isMuted = false,
     this.volume = 1.0,
     this.errorDescription,
   });
@@ -63,6 +64,9 @@ class VideoPlayerValue {
 
   /// True if the video is currently buffering.
   final bool isBuffering;
+
+  /// True if the video is muted.
+  final bool isMuted;
 
   /// The current volume of the playback.
   final double volume;
@@ -98,6 +102,7 @@ class VideoPlayerValue {
     bool isPlaying,
     bool isLooping,
     bool isBuffering,
+    bool isMuted,
     double volume,
     String errorDescription,
   }) {
@@ -109,6 +114,7 @@ class VideoPlayerValue {
       isPlaying: isPlaying ?? this.isPlaying,
       isLooping: isLooping ?? this.isLooping,
       isBuffering: isBuffering ?? this.isBuffering,
+      isMuted: isMuted ?? this.isMuted,
       volume: volume ?? this.volume,
       errorDescription: errorDescription ?? this.errorDescription,
     );
@@ -123,7 +129,8 @@ class VideoPlayerValue {
         'buffered: [${buffered.join(', ')}], '
         'isPlaying: $isPlaying, '
         'isLooping: $isLooping, '
-        'isBuffering: $isBuffering'
+        'isBuffering: $isBuffering, '
+        'isMuted: $isMuted, '
         'volume: $volume, '
         'errorDescription: $errorDescription)';
   }
@@ -313,6 +320,13 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     await _applyLooping();
   }
 
+  /// Sets whether or not video should be muted. See also
+  /// [VideoPlayerValue.isMuted].
+  Future<void> setMuted(bool muted) async {
+    value = value.copyWith(isMuted: muted);
+    await _applyMuted();
+  }
+
   /// Pauses the video.
   Future<void> pause() async {
     value = value.copyWith(isPlaying: false);
@@ -324,6 +338,13 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
       return;
     }
     await VideoPlayerPlatform.instance.setLooping(_textureId, value.isLooping);
+  }
+
+  Future<void> _applyMuted() async {
+    if (!value.initialized || _isDisposed) {
+      return;
+    }
+    await VideoPlayerPlatform.instance.setMuted(_textureId, value.isMuted);
   }
 
   Future<void> _applyPlayPause() async {