+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.6
+
+* Add `muted` attribute
+
 ## 0.10.5
 
 * Support `web` by default.