+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.5
+
+* Add `muted` attribute
+
 ## 1.0.4
 
 * Remove the deprecated `author:` field from pubspec.yaml