+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+* Add `muted` attribute
+
 ## 0.1.1+1
 
 - Add an android/ folder with no-op implementation to workaround https://github.com/flutter/flutter/issues/46898.